Update README.md (#15)
Fixed headings to actually appear with proper formatting, and not just as pound symbols. Also moved the CircleCI link (it isn't working for me, but I don't know if that's just because I don't have a CircleCI account) lower in the page so it's not stuck weirdly between two images.